Add golangci-lint configuration and update Makefile
- Introduced .golangci.yml for linting configuration.
- Added `lint` and `fmt` targets to the Makefile.
- Enabled race detection in the `build-prod` target.
- Updated comments and error messages for consistency.
- Refactored code to use `errors.New` for static error strings.
- Improved logging and formatting across multiple files.

+++ client/internal/flagparser/parse.go
@@ -23,7 +23,7 @@ func ParseLine(line string) (models.Flag, string, error) {
 		return models.Flag{}, out.Status, errors.New(out.Message)
 	case "failed":
 		return models.Flag{}, out.Status, fmt.Errorf("flag submission failed for team %d on the %s: %s",
-			out.TeamID, utils.MapPortToService(uint16(out.PortService)), out.Message)
+			out.TeamID, utils.MapPortToService(out.PortService), out.Message)
 	case "error":
 		return models.Flag{}, out.Status, fmt.Errorf("flag submission error: %s", out.Message)
 	case "fatal":
@@ -31,7 +31,7 @@ func ParseLine(line string) (models.Flag, string, error) {
 	case "success":
 		return models.Flag{
 			FlagCode:     out.FlagCode,
-			ServiceName:  utils.MapPortToService(uint16(out.PortService)),
+			ServiceName:  utils.MapPortToService(out.PortService),
 			PortService:  out.PortService,
 			SubmitTime:   uint64(time.Now().Unix()),
 			ResponseTime: 0,
